Auth frontend
Summary:

Test Plan:
