+++ packages/app/entrypoints/EntrypointRegister.tsx
@@ -2,7 +2,7 @@ import { iso } from "packages/app/__generated__/__isograph/iso.ts";
 import type { RouteEntrypoint } from "packages/app/routes.ts";
 
 export const EntrypointRegister = iso(`
-  field Query.EntrypointRegister($code: ID!) {
+  field Query.EntrypointRegister($code: String!) {
     RegistrationForm(code: $code)
   }
 `)(function EntrypointRegister({ data }): RouteEntrypoint {
